fix(map-calibration): persist physical capture rect at snip time, drop per-monitor DPI map (#947)
Review fix: the read-time DIU→physical conversion was wrong on mixed-DPI
multi-monitor and over-engineered. The snip (RegionSnipWindow) is a SINGLE
WPF window spanning the whole virtual desktop, so under PerMonitorV2 it has
ONE DPI scale and maps its entire logical surface uniformly — WPF does not
per-monitor-rescale GetPosition within one window. The provider's
MonitorDpiProvider + MonitorScaleSelector reconstructed a per-monitor DIU
layout and applied a per-monitor affine map, a frame the snip never
produces; on a 1920@100% + secondary@150% layout it computed origins off by
the secondary's full physical width. Its mixed-DPI test was tautological
(hand-fed a MonitorDpiInfo MonitorDpiProvider cannot emit), so green CI was
illusory.

Fix: persist PHYSICAL desktop pixels computed once at snip-confirm time from
the snip window's own live TransformToDevice scale, so the read path is
frame-independent.

- IMapCaptureRectStore now deals in CaptureRect (int physical px); the
  MapCaptureRectDiu record is deleted. ShellSettings.MapCaptureBbox stores
  int Left/Top/Width/Height (still nullable + additive, no migration).
- RegionSnipWindow exposes BOTH SelectedRect (DIU, for the overlay mirror)
  and SelectedPhysicalRect (physical, via CaptureRectMath.DiuToPhysical with
  the window's single transform scale). MapBboxDrawController's snip seam
  carries both (internal SnipResult record); it persists the physical rect
  and still mirrors the DIU rect onto the overlay.
- MapCaptureRegionProvider.Current = store?.Get() returned verbatim (null on
  unset/degenerate). Removed IMonitorDpiProvider/MonitorDpiProvider/
  MonitorScaleSelector/MonitorDpiInfo, the affine map, the per-monitor DPI
  P/Invokes in NativeMethods.txt (EnumDisplayMonitors/GetMonitorInfo/
  GetDpiForMonitor/MONITOR_DPI_TYPE), the DI registration, and the CA1416/
  OS-guard complexity.
- Tests: deleted the tautological mixed-DPI affine test and the
  MonitorScaleSelector/MonitorDpiProvider tests; added snip DIU→physical
  coverage at 1.5x (incl. negative virtual origin); kept/adjusted the store
  round-trip + provider fail-soft tests to the physical-px shape.
- Documented the known limitation (single-monitor + uniform-DPI correct;
  mixed-DPI multi-monitor owed to #938 manual-verify; stale on later
  DPI/resolution change → re-snip) in RegionSnipWindow + IMapCaptureRectStore.

Decoder-free (#921) preserved. One-rect invariant intact (capture frame ==
overlay frame at snip time). Capture + Shell + MapCalibration test projects
and ShippedGraphDecoderFreeTests all green.

+++ src/Mithril.Shell/ShellSettings.cs
@@ -78,7 +78,9 @@ public sealed class ShellSettings : INotifyPropertyChanged, IActiveCharacterPers
     public double SidebarWidth { get => _sidebarWidth; set => Set(ref _sidebarWidth, value); }
 
     /// <summary>
-    /// The persisted map-capture bbox (#947), in absolute virtual-desktop DIUs.
+    /// The persisted map-capture bbox (#947), in absolute virtual-desktop PHYSICAL
+    /// pixels (the frame BitBlt reads — resolved at snip-confirm time from the snip
+    /// window's own device scale, so the read path needs no DPI work).
     /// <see langword="null"/> until the user first snips a region (the legitimate
     /// "no bbox set" state). Owned by the shell — not Legolas — because the shell
     /// references both Legolas and the Capture project, so a shell-side
@@ -109,21 +111,22 @@ public sealed class ShellSettings : INotifyPropertyChanged, IActiveCharacterPers
 }
 
 /// <summary>
-/// The persisted map-capture bbox (#947), in absolute virtual-desktop DIUs (1 DIU =
-/// 1/96"). A plain POCO so STJ source-gen serializes it as a nested object on
-/// <see cref="ShellSettings.MapCaptureBbox"/>. The DIU frame matches WPF
-/// <c>Window.Left/Top</c> and the snip's absolute virtual-desktop output; the
-/// Capture project converts it to physical pixels at read time via the live
-/// per-monitor DPI layout. Mirrors the Capture project's <c>MapCaptureRectDiu</c>
-/// (kept separate so the shell-settings schema doesn't take a Capture dependency in
-/// its persisted shape).
+/// The persisted map-capture bbox (#947), in absolute virtual-desktop PHYSICAL
+/// pixels (the frame <c>GetDC(NULL)</c>/<c>BitBlt</c> read; origin signed on a
+/// multi-monitor layout). A plain POCO so STJ source-gen serializes it as a nested
+/// object on <see cref="ShellSettings.MapCaptureBbox"/>. The physical rect is
+/// resolved once at snip-confirm time from the snip window's single device scale, so
+/// the Capture project's region provider returns it verbatim with no read-time DPI
+/// work. Mirrors the Capture project's <c>CaptureRect</c> (kept as a separate POCO
+/// so the shell-settings schema doesn't take a Capture dependency in its persisted
+/// shape).
 /// </summary>
 public sealed class MapCaptureBbox
 {
-    public double Left { get; set; }
-    public double Top { get; set; }
-    public double Width { get; set; }
-    public double Height { get; set; }
+    public int Left { get; set; }
+    public int Top { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
 }
 
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase, WriteIndented = true)]
